pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openolympus"
version = "0.1.0"
description = "OpenOlympus — AI Multi-Agent Collaboration Framework"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "aiofiles>=24.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
